# build/resources/main/openapi/api.yaml
openapi: 3.0.3
info:
  title: Sheets API
  description: API for collaborative spreadsheet application
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Proprietary
servers:
  - url: /v1
    description: Development server
paths:
...
  /sheet/{sheetId}/cell:
    post:
      summary: Update a cell
      description: Updates an existing cell in a sheet. Cells are automatically created when a sheet is created.
      operationId: updateCell
      tags:
        - Cell
      parameters:
        - name: sheetId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the sheet
        - name: X-User-ID
          in: header
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user making the request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CellRequest'
      responses:
        '200':
          description: Cell updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellErrorResponse'
        '403':
          description: Forbidden - User does not have write access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sheet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Cell is locked by another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    get:
      summary: Get all cells in a sheet
      description: Retrieves all cells in a sheet
      operationId: getCellsBySheetId
      tags:
        - Cell
      parameters:
        - name: sheetId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the sheet
        - name: X-User-ID
          in: header
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user making the request
      responses:
        '200':
          description: Cells retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellBatchResponse'
        '403':
          description: Forbidden - User does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sheet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sheet/{sheetId}/cell/{row}/{column}:
    get:
      summary: Get a specific cell
      description: Retrieves a specific cell from a sheet
      operationId: getCell
      tags:
        - Cell
      parameters:
        - name: sheetId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the sheet
        - name: row
          in: path
          required: true
          schema:
            type: integer
          description: Row index of the cell
        - name: column
          in: path
          required: true
          schema:
            type: integer
          description: Column index of the cell
        - name: X-User-ID
          in: header
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user making the request
      responses:
        '200':
          description: Cell retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellResponse'
        '403':
          description: Forbidden - User does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sheet or cell not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    delete:
      summary: Delete a cell
      description: Deletes a specific cell from a sheet
      operationId: deleteCell
      tags:
        - Cell
      parameters:
        - name: sheetId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the sheet
        - name: row
          in: path
          required: true
          schema:
            type: integer
          description: Row index of the cell
        - name: column
          in: path
          required: true
          schema:
            type: integer
          description: Column index of the cell
        - name: X-User-ID
          in: header
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user making the request
      responses:
        '204':
          description: Cell deleted successfully
        '403':
          description: Forbidden - User does not have write access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sheet or cell not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Cell is locked by another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sheet/{sheetId}/cell/{cellId}:
    post:
      summary: Update a cell by its ID
      description: Updates an existing cell in a sheet using its direct cell ID.
      operationId: updateCellById
      tags:
        - Cell
      parameters:
        - name: sheetId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the sheet
        - name: cellId
          in: path
          required: true
          schema:
            type: string
          description: ID of the cell to update (format sheetId:row:column)
        - name: X-User-ID
          in: header
          required: true
          schema:
            type: integer
            format: int64
          description: ID of the user making the request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CellDataRequest'
      responses:
        '200':
          description: Cell updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellErrorResponse'
        '403':
          description: Forbidden - User does not have write access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Sheet or cell not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Cell is locked by another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Error message
          example: "Invalid request parameters"
        path:
          type: string
          description: Request path
          example: "/v1/sheet"
        timestamp:
          type: string
          format: date-time
          description: Time when the error occurred
          example: "2025-05-24T12:00:00Z"
    CellRequest:
      $ref: 'schemas/cell.yaml#/components/schemas/CellRequest'
    CellResponse:
      $ref: 'schemas/cell.yaml#/components/schemas/CellResponse'
...
    CellBatchResponse:
      $ref: 'schemas/cell.yaml#/components/schemas/CellBatchResponse'
    CellErrorResponse:
      $ref: 'schemas/cell.yaml#/components/schemas/CellErrorResponse'
    CellDataRequest:
      type: object
      properties:
        value:
          type: string
          description: New value of the cell
          example: "Hello, World!"