# src/main/kotlin/com/sheets/controllers/CellController.kt
package com.sheets.controllers

import com.sheets.generated.api.CellApi
import com.sheets.generated.model.CellBatchResponse
import com.sheets.generated.model.CellDataRequest
import com.sheets.generated.model.CellRequest
import com.sheets.generated.model.CellResponse
import com.sheets.models.domain.Cell
import com.sheets.models.domain.DataType
import com.sheets.services.CellService
import com.sheets.services.SheetService
import com.sheets.services.expression.exception.CircularDependencyException
import com.sheets.services.expression.exception.ExpressionException
import org.slf4j.LoggerFactory
import org.springframework.http.HttpStatus
import org.springframework.http.ResponseEntity
import org.springframework.web.bind.annotation.RestController
import java.time.Instant
import java.time.OffsetDateTime
import java.time.ZoneId

@RestController
class CellController(
    private val cellService: CellService,
    private val sheetService: SheetService
) : CellApi {

    private val logger = LoggerFactory.getLogger(CellController::class.java)

    override fun updateCell(
        sheetId: Long,
        xUserID: Long,
        cellRequest: CellRequest
    ): ResponseEntity<CellResponse> {
        logger.info("Received request to update cell at row: {}, column: {} in sheet: {} by user: {}", 
            cellRequest.row, cellRequest.column, sheetId, xUserID)
        
        try {
            // Validate that the sheet exists and user has access
            logger.debug("Validating sheet access for sheet: {} and user: {}", sheetId, xUserID)
            sheetService.getSheetById(sheetId, xUserID)

            // Create a cell object from the request
            val cellId = "${sheetId}:${cellRequest.row}:${cellRequest.column}"
            logger.debug("Cell ID for update: {}", cellId)
            
            try {
                // Get existing cell
                logger.debug("Retrieving existing cell with ID: {}", cellId)
                val existingCell = cellService.getCell(cellId) ?: run {
                    logger.warn("Cell not found with ID: {}", cellId)
                    return ResponseEntity
                        .status(HttpStatus.NOT_FOUND)
                        .body(null)
                }
                    
                // Update the cell
                logger.debug("Updating cell with ID: {}", cellId)
                val updatedCell = cellService.updateCell(
                    existingCell.copy(
                        data = cellRequest.data
                    ),
                    xUserID.toString()
                )
                
                // Convert to response
                val response = toCellResponse(updatedCell)
                logger.info("Successfully updated cell at row: {}, column: {} in sheet: {} by user: {}", 
                    cellRequest.row, cellRequest.column, sheetId, xUserID)
                
                return ResponseEntity.ok(response)
            } catch (e: CircularDependencyException) {
                logger.error("Circular dependency detected while updating cell at row: {}, column: {} in sheet: {} by user: {}: {}", 
                    cellRequest.row, cellRequest.column, sheetId, xUserID, e.message)
                return ResponseEntity
                    .status(HttpStatus.BAD_REQUEST)
                    .body(null)
            } catch (e: ExpressionException) {
                logger.error("Expression error while updating cell at row: {}, column: {} in sheet: {} by user: {}: {}", 
                    cellRequest.row, cellRequest.column, sheetId, xUserID, e.message)
                return ResponseEntity
                    .status(HttpStatus.BAD_REQUEST)
                    .body(null)
            } catch (e: IllegalStateException) {
                // This is likely a locking issue
                logger.error("Lock acquisition failed while updating cell at row: {}, column: {} in sheet: {} by user: {}: {}", 
                    cellRequest.row, cellRequest.column, sheetId, xUserID, e.message)
                return ResponseEntity
                    .status(HttpStatus.CONFLICT)
                    .body(null)
            } catch (e: Exception) {
                logger.error("Unexpected error while updating cell at row: {}, column: {} in sheet: {} by user: {}: {}", 
                    cellRequest.row, cellRequest.column, sheetId, xUserID, e.message, e)
                return ResponseEntity
                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(null)
            }
        } catch (e: Exception) {
            logger.error("Error validating sheet access for sheet: {} and user: {}: {}", 
                sheetId, xUserID, e.message, e)
            return ResponseEntity
                .status(HttpStatus.INTERNAL_SERVER_ERROR)
                .body(null)
        }
    }
    

    /**
     * Updates a cell by its direct ID
     */
    override fun updateCellById(
        sheetId: Long,
        cellId: String,
        xUserID: Long,
        cellDataRequest: CellDataRequest
    ): ResponseEntity<CellResponse> {
        logger.info("Received request to update cell with ID: {} in sheet: {} by user: {}", 
            cellId, sheetId, xUserID)
        
        try {
            // Validate that the sheet exists and user has access
            logger.debug("Validating sheet access for sheet: {} and user: {}", sheetId, xUserID)
            sheetService.getSheetById(sheetId, xUserID)
            
            // Validate that the cell ID belongs to the sheet
            if (!cellId.startsWith("$sheetId:")) {
                logger.warn("Cell ID: {} does not belong to sheet: {}", cellId, sheetId)
                return ResponseEntity
                    .status(HttpStatus.BAD_REQUEST)
                    .body(null)
            }
            
            // Validate that the value is not null
            val cellValue = cellDataRequest.value
            if (cellValue == null) {
                logger.warn("Cell value cannot be null for cell ID: {}", cellId)
                return ResponseEntity
                    .status(HttpStatus.BAD_REQUEST)
                    .body(null)
            }
            
            try {
                // Get existing cell
                logger.debug("Retrieving existing cell with ID: {}", cellId)
                val existingCell = cellService.getCell(cellId) ?: run {
                    logger.warn("Cell not found with ID: {}", cellId)
                    return ResponseEntity
                        .status(HttpStatus.NOT_FOUND)
                        .body(null)
                }
                    
                // Update the cell
                logger.debug("Updating cell with ID: {}", cellId)
                val updatedCell = cellService.updateCell(
                    existingCell.copy(
                        data = cellValue
                    ),
                    xUserID.toString()
                )
                
                // Convert to response
                val response = toCellResponse(updatedCell)
                logger.info("Successfully updated cell with ID: {} in sheet: {} by user: {}", 
                    cellId, sheetId, xUserID)
                
                return ResponseEntity.ok(response)
            } catch (e: CircularDependencyException) {
                logger.error("Circular dependency detected while updating cell with ID: {} in sheet: {} by user: {}: {}", 
                    cellId, sheetId, xUserID, e.message)
                return ResponseEntity
                    .status(HttpStatus.BAD_REQUEST)
                    .body(null)
            } catch (e: ExpressionException) {
                logger.error("Expression error while updating cell with ID: {} in sheet: {} by user: {}: {}", 
                    cellId, sheetId, xUserID, e.message)
                return ResponseEntity
                    .status(HttpStatus.BAD_REQUEST)
                    .body(null)
            } catch (e: IllegalStateException) {
                // This is likely a locking issue
                logger.error("Lock acquisition failed while updating cell with ID: {} in sheet: {} by user: {}: {}", 
                    cellId, sheetId, xUserID, e.message)
                return ResponseEntity
                    .status(HttpStatus.CONFLICT)
                    .body(null)
            } catch (e: Exception) {
                logger.error("Unexpected error while updating cell with ID: {} in sheet: {} by user: {}: {}", 
                    cellId, sheetId, xUserID, e.message, e)
                return ResponseEntity
                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .body(null)
            }
        } catch (e: Exception) {
            logger.error("Error validating sheet access for sheet: {} and user: {}: {}", 
                sheetId, xUserID, e.message, e)
            return ResponseEntity
                .status(HttpStatus.INTERNAL_SERVER_ERROR)
                .body(null)
        }
    }
    
    override fun getCellsBySheetId(
        sheetId: Long,
        xUserID: Long
    ): ResponseEntity<CellBatchResponse> {
        logger.info("Received request to get all cells in sheet: {} by user: {}", sheetId, xUserID)
        
        try {
            // Validate that the sheet exists and user has access
            logger.debug("Validating sheet access for sheet: {} and user: {}", sheetId, xUserID)
            sheetService.getSheetById(sheetId, xUserID)
                
            // Get cells
            logger.debug("Retrieving cells for sheet: {}", sheetId)
            val cells = cellService.getCellsBySheetId(sheetId)
            
            // Convert to response
            val cellResponses = cells.map { toCellResponse(it) }
            logger.info("Successfully retrieved {} cells from sheet: {} for user: {}", 
                cells.size, sheetId, xUserID)
            
            return ResponseEntity.ok(CellBatchResponse(cellResponses))
        } catch (e: Exception) {
            logger.error("Error retrieving cells for sheet: {} and user: {}: {}", 
                sheetId, xUserID, e.message, e)
            return ResponseEntity
                .status(HttpStatus.INTERNAL_SERVER_ERROR)
                .body(null)
        }
    }
    
    override fun getCell(
        sheetId: Long,
        row: Int,
        column: Int,
        xUserID: Long
    ): ResponseEntity<CellResponse> {
        logger.info("Received request to get cell at row: {}, column: {} in sheet: {} by user: {}", 
            row, column, sheetId, xUserID)
        
        try {
            // Validate that the sheet exists and user has access
            logger.info("Validating sheet access for sheet: {} and user: {}", sheetId, xUserID)
            sheetService.getSheetById(sheetId, xUserID)
                
            // Get cell
            val cellId = "${sheetId}:${row}:${column}"
            logger.debug("Retrieving cell with ID: {}", cellId)
            val cell = cellService.getCell(cellId) ?: run {
                logger.info("Cell not found with ID: {}", cellId)
                return ResponseEntity.notFound().build()
            }
                
            // Convert to response
            val response = toCellResponse(cell)
            logger.info("Successfully retrieved cell at row: {}, column: {} in sheet: {} for user: {}", 
                row, column, sheetId, xUserID)
            
            return ResponseEntity.ok(response)
        } catch (e: Exception) {
            logger.error("Error retrieving cell at row: {}, column: {} in sheet: {} for user: {}: {}", 
                row, column, sheetId, xUserID, e.message, e)
            return ResponseEntity
                .status(HttpStatus.INTERNAL_SERVER_ERROR)
                .body(null)
        }
    }
    
    override fun deleteCell(
        sheetId: Long,
        row: Int,
        column: Int,
        xUserID: Long
    ): ResponseEntity<Unit> {
        logger.info("Received request to delete cell at row: {}, column: {} in sheet: {} by user: {}", 
            row, column, sheetId, xUserID)
        
        try {
            // Validate that the sheet exists and user has access
            logger.debug("Validating sheet access for sheet: {} and user: {}", sheetId, xUserID)
            sheetService.getSheetById(sheetId, xUserID)
                
            // Delete cell
            val cellId = "${sheetId}:${row}:${column}"
            logger.debug("Deleting cell with ID: {}", cellId)
            
            try {
                cellService.deleteCell(cellId, xUserID.toString())
                logger.info("Successfully deleted cell at row: {}, column: {} in sheet: {} by user: {}", 
                    row, column, sheetId, xUserID)
                return ResponseEntity.noContent().build()
            } catch (e: IllegalStateException) {
                // This is likely a locking issue
                logger.error("Lock acquisition failed while deleting cell at row: {}, column: {} in sheet: {} by user: {}: {}", 
                    row, column, sheetId, xUserID, e.message)
                return ResponseEntity
                    .status(HttpStatus.CONFLICT)
                    .build()
            } catch (e: Exception) {
                logger.error("Unexpected error while deleting cell at row: {}, column: {} in sheet: {} by user: {}: {}", 
                    row, column, sheetId, xUserID, e.message, e)
                return ResponseEntity
                    .status(HttpStatus.INTERNAL_SERVER_ERROR)
                    .build()
            }
        } catch (e: Exception) {
            logger.error("Error validating sheet access for sheet: {} and user: {}: {}", 
                sheetId, xUserID, e.message, e)
            return ResponseEntity
                .status(HttpStatus.INTERNAL_SERVER_ERROR)
                .build()
        }
    }
    
    private fun toCellResponse(cell: Cell): CellResponse {
        logger.debug("Converting domain Cell to CellResponse: {}", cell.id)
        return CellResponse(
            id = cell.id,
            sheetId = cell.sheetId,
            row = cell.row,
            column = cell.column,
            data = cell.data,
            dataType = when (cell.dataType) {
                DataType.PRIMITIVE -> CellResponse.DataType.PRIMITIVE
                DataType.EXPRESSION -> CellResponse.DataType.EXPRESSION
            },
            evaluatedValue = cell.evaluatedValue,
            createdAt = cell.createdAt.atZone(ZoneId.systemDefault()).toOffsetDateTime(),
            updatedAt = cell.updatedAt.atZone(ZoneId.systemDefault()).toOffsetDateTime()
        )
    }
}